+++ src/IDBFactory.js
@@ -1,6 +1,6 @@
 import {createEvent, ShimEvent, ProxyPolyfill, IDBVersionChangeEvent} from './Event.js';
 import {findError, createDOMException, DOMException} from './DOMException.js';
-import {IDBOpenDBRequest} from './IDBRequest.js';
+import {IDBOpenDBRequest, IDBRequest} from './IDBRequest.js';
 import * as util from './util.js';
 import Key from './Key.js';
 import IDBTransaction from './IDBTransaction.js';
@@ -86,7 +86,6 @@ IDBFactory.prototype.open = function (name, version) {
         db.transaction(function (tx) {
             tx.executeSql('CREATE TABLE IF NOT EXISTS __sys__ (name VARCHAR(255), keyPath VARCHAR(255), autoInc BOOLEAN, indexList BLOB, currNum INTEGER)', [], function () {
                 tx.executeSql('SELECT * FROM __sys__', [], function (tx, data) {
-                    const e = createEvent('success');
                     req.__result = new IDBDatabase(db, name, version, data);
                     if (oldVersion < version) {
                         // DB Upgrade in progress
@@ -112,6 +111,7 @@ IDBFactory.prototype.open = function (name, version) {
                             }, dbCreateError);
                         }, dbCreateError);
                     } else {
+                        const e = createEvent('success');
                         req.dispatchEvent(e);
                     }
                 }, dbCreateError);
@@ -259,6 +259,40 @@ function cmp (key1, key2) {
 }
 
 IDBFactory.prototype.cmp = cmp;
+
+/**
+* NON-STANDARD!! (Also may return outdated information)
+* @link https://www.w3.org/Bugs/Public/show_bug.cgi?id=16137
+* @link http://lists.w3.org/Archives/Public/public-webapps/2011JulSep/1537.html
+*/
+IDBFactory.prototype.webkitGetDatabaseNames = function () {
+    let calledDbCreateError = false;
+    function dbGetDatabaseNamesError (...args /* tx, err */) {
+        if (calledDbCreateError) {
+            return;
+        }
+        const err = findError(args);
+        calledDbCreateError = true;
+        const evt = createEvent('error', args);
+        req.__readyState = 'done';
+        req.__error = err || DOMException;
+        req.dispatchEvent(evt);
+    }
+    const req = new IDBRequest();
+    createSysDB(function () {
+        sysdb.transaction(function (tx) {
+            tx.executeSql('SELECT name FROM dbVersions', [], function (tx, data) {
+                const dbNames = new util.StringList();
+                for (let i = 0; i < data.rows.length; i++) {
+                    dbNames.push(data.rows.item(i).name);
+                }
+            }, dbGetDatabaseNamesError);
+        }, dbGetDatabaseNamesError);
+    }, dbGetDatabaseNamesError);
+    const e = createEvent('success');
+    req.dispatchEvent(e);
+    return req;
+};
 
 IDBFactory.prototype.toString = function () {
     return '[object IDBFactory]';